Fixed some bugs related to the uniqueness of the node slug
In some queries in the functions for editing and saving nodes several items were returned to get() - which leads to an error - when the same node name was used across several trees. Fixed by including an additional filter to make sure that only nodes of the current tree are queried.